Adjusted parameter lookup to fix length of text type. (#38)

+++ DataAccessGeneration/ResultDefinition.cs
@@ -5,7 +5,7 @@
 		public string Name { get; set; } = "";
 		public string TypeName { get; set; } = "";
 		public bool? IsNullable { get; set; }
-		public short? MaxLength { get; set; }
+		public int? MaxLength { get; set; }
 		public byte? Precision { get; set; }
 		public byte? Scale { get; set; }
 	}
